// src/api/models/ServiceAttributeModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ServiceAttributeModel = {
  name: string;
  code: string;
  description?: string | null;
  dataType?: string | null;
  isRequired?: boolean;
  isActive?: boolean;
  sortOrder?: number;
  options?: Array<string> | null;
  validationRules?: Record<string, any> | null;
  businessCategoryId: string;
};