// src/api/models/ServiceAttributeViewStandardResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HttpStatusCode } from './HttpStatusCode';
import type { ServiceAttributeView } from './ServiceAttributeView';
export type ServiceAttributeViewStandardResponse = {
  href?: string | null;
  relations?: Array<string> | null;
  method?: string | null;
  routeName?: string | null;
  routeValues?: any;
  status?: boolean;
  message?: string | null;
  data?: ServiceAttributeView;
  statusCode?: HttpStatusCode;
  errors?: any;
};