// src/api/models/ServiceCategoryModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ServiceCategoryModel = {
  name: string;
  code: string;
  description?: string | null;
  icon?: string | null;
  isActive?: boolean;
  sortOrder?: number;
  parentCategoryId?: string | null;
  businessCategoryId: string;
};