// src/api/models/ServiceTemplateViewStandardResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HttpStatusCode } from './HttpStatusCode';
import type { ServiceTemplateView } from './ServiceTemplateView';
export type ServiceTemplateViewStandardResponse = {
  href?: string | null;
  relations?: Array<string> | null;
  method?: string | null;
  routeName?: string | null;
  routeValues?: any;
  status?: boolean;
  message?: string | null;
  data?: ServiceTemplateView;
  statusCode?: HttpStatusCode;
  errors?: any;
};